CHanges
Changing the main engine sheet so that it is much more simplified and
streamlined and you click to expand the individual panels. Next step
will be to add the aux engine stuff to the same sheet, as well as the
fuel inputs. So we have one sheet for input, and another sheet that is
just straight results.